Re-org request and response code in separate directories
