+++ packages/lodestar/src/network/reqresp/request/ttfbTimeoutController.ts
@@ -1,6 +1,6 @@
 import {AbortController} from "abort-controller";
 import {source as abortSource} from "abortable-iterator";
-import {TtfbTimeoutError} from "../errors";
+import {ResponseErrorCode, ResponseInternalError} from "./errors";
 
 const abortMessage = "TTFB_TIMEOUT_ERROR_MESSAGE";
 
@@ -31,12 +31,14 @@ export function ttfbTimeoutController(
     } catch (e) {
       // Rethrow error properly typed so the peer score can pick it up
       if (e.message === abortMessage) {
-        throw new TtfbTimeoutError();
+        throw new ResponseInternalError({code: ResponseErrorCode.TTFB_TIMEOUT});
       } else {
         throw e;
       }
     } finally {
-      if (responseTimer) clearTimeout(responseTimer);
+      if (responseTimer) {
+        clearTimeout(responseTimer);
+      }
     }
   };
 }
